Explicitly define CMAKE_INSTALL_CMAKEDIR to work around FTBFS for now.
Calamares uses plain GNUInstallDirs, but includes a submodule based on
kde-partitionmanager which uses ECM (extra-cmake-modules). With
extra-cmake-modules 1.6.0, there appears to be some bad interaction
which leads to CMAKE_INSTALL_LIBDIR being empty on i686 at the line
of CMakeLists.txt where CMAKE_INSTALL_CMAKEDIR is defined in terms of
it.